Fix Default Color Schemes in Light & Dark Mode
Hi! I love this theme, it's very beautiful!

One thing I noticed is that upon installation, the colors are simple desaturated Obsidian values by default. That is because the theme relies on Style Settings and CSS classes for its colors. You need to install Style Settings and choose the default color schemes for it to work.

These changes make it so that by default, even without these Style Settings classes, the colors match the theme. So when first installed, without these classes, the colors work as intended, but when a different class is selected from Style Settings it will still work as usual.

Also, thanks for making this theme! Keep up the great work.